+++ app/src/main/java/app/urv/manager/di/DatabaseModule.kt
@@ -20,6 +20,7 @@ import app.urv.manager.data.room.MIGRATION_14_15
 import app.urv.manager.data.room.MIGRATION_15_16
 import app.urv.manager.data.room.MIGRATION_16_17
 import app.urv.manager.data.room.MIGRATION_17_18
+import app.urv.manager.data.room.MIGRATION_18_19
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
@@ -43,7 +44,8 @@ val databaseModule = module {
                 MIGRATION_14_15,
                 MIGRATION_15_16,
                 MIGRATION_16_17,
-                MIGRATION_17_18
+                MIGRATION_17_18,
+                MIGRATION_18_19
             )
             .build()